mpegmap/TC_5.1.2-1_VideoTrack.html

<!doctype html>
<html>
  <head>
    <title>TC 5.1.2-1 VideoTrack</title>
    <link rel="author" title="CableLabs" href="html5@cablelabs.com">
    <meta name="flags" content="[requirement flags]">
    <meta name="assert" content="MPEG-2 SPTS - UA creates VideoTrack, attributes set.">
    <meta name="timeout" content="long">
    <script src="testmedia.js"></script>
    <!-- Source HTTP server local copy of World Wide Web Consortium (W3C) Test Harness. --> 
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <!-- Script Version: BETA -->

    <!-- About the Media Resource
           To verify DOM elements the details of the test media must be known prior to test execution.
           This test script is bound to the test media file sourced in the script.
           Sourcing a differnt test media file, or modifying that file, may invalidate the test.

           The media is an MPEG-2, Single Program Transport Stream
           containing one video elementary stream, ES PID 33.
    -->
  </head>


  <body>
    <p><a href="http://html5.cablelabs.com">spec reference</a></p>
    <p><a href="http://html5.cablelabs.com/dlna-rui-test/dlna">test assertion</a></p>
    <video id="v" width="400" height="300" autoplay controls></video>
    <div id="log"></div>


    <script>
      var test = async_test("VideoTrack attributes.");
      var vid = document.getElementById("v");
      vid.src = mediaServerURL() + "UserPrivateStreams.ts";

      var index = -1;
      var pid = '33'; // video ES PID
      var kind = 'main';


      vid.addEventListener("playing", function() {

        vid.pause();

        setTimeout(function() {
          for (i=0; i< vid.videoTracks.length; i++) {
            if (vid.videoTracks[i].id == pid) {
              index = i;
              break;
            }
          }
          if (index == -1) {
            test.step(function() {
              assert_unreached("Did not find VideoTrack PID " + pid + " in DOM");
            }, "Did not find VideoTrack " + pid + " in DOM");
            test.done();

          } else {
            test.step(function() {
              assert_equals(vid.videoTracks.length, 1, "VideoTracks.length");
              assert_equals(vid.videoTracks[index].id, pid, "VideoTrack.id");
              assert_equals(vid.videoTracks[index].kind, kind, "VideoTrack.kind");
            });
            test.done();
          }
        }, 12000);
      }, false);

    </script>
  </body>
</html>
